d3aa: got weak battery detection actually working, and not letting the magic smoke out of updi adapters any more (probably)
The alkaline detection might be a little too lenient though; it
could potentially fail to activate limits when the cell is completely
full or stronger than an average alkaline.  One of my test cells
measured at 72 / 75, so if it was just a little stronger it'd pass...
but most alkalines I tried were in the 40 to 60 range and failed easily.

OTOH, if I make it easier to fail, it's likely to trip on normal li-ion
cells, and I don't want that.

So as a future enhancement idea, maybe it should have a smaller sag
threshold for AA and a larger threshold for li-ion.  That would reduce
false negatives for AA, while still preventing false positives for li-ion.